requirements.txt
python-dotenv==1.0.1
squareup==37.1.0.20240604